Gan/requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
Pillow>=9.5.0
tqdm>=4.65.0

# Optional but recommended
matplotlib>=3.7.0  # For visualization
scikit-learn>=1.2.0  # For additional metrics
pandas>=2.0.0  # For data analysis

# Development
pytest>=7.3.0  # For testing
black>=23.0.0  # For code formatting